requirements.txt
fastapi[standard]==0.115.14
chromadb~=1.0.13
requests~=2.32.4
sentence-transformers~=5.0.0
pillow~=11.3.0
open-clip-torch~=2.32.0
pytest~=8.0.0
pytest-mock~=3.12.0
pytest-cov~=4.0.0
flake8~=7.3.0
flake8-bugbear~=24.12.12
numpy~=2.3.2
python-multipart~=0.0.20